Move settings test to its own file
